i3: solve the Compose window scratchpad problem with Thunderbird
As per the previous configuration, any window with the Thunderbird
class was being moved to the scratchpad workspace whenever it was
created. This caused the Compose Mail window also to behave the same
way, which was annoying since I had to press Ctrl+N and Mod+T every
time to get that window back.

This resolves that by putting tighter constraints on which Thunderbird
window will move to the scratchpad workspace.