config: revamp init-keyboard script
The ideas are primarily taken from here:
https://gist.github.com/joshskidmore/f8599610da3308613cddcaf46a701ecc

Instead of the complications of running xcape as a service and setting
up xmodmap using the configuration files, we simply directly call the
xmodmap and xcape command line applications with the configurations
that we want. I find this much easier to parse and maintain in the
long run. The intent is conveyed in a clearer fashion.